api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tnww-api"
version = "0.1.0"
description = "TNWW v3 API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "pandas>=2.2.0",
  "numpy>=2.0.0",
  "PyYAML>=6.0.0",
  "APScheduler>=3.10.0",
  "openpyxl>=3.1.0",
  "python-dateutil>=2.9.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
